bot/requirements.txt
#Basic
arrow==0.15.5

# discord
discord.py

# Database
psycopg2-binary==2.8.5

# Docs
drf-yasg==1.17.1

# Linters
pylint==2.5.0
vulture==2.1

# Tests
pytest==6.0.0
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3
pytest-lazy-fixture==0.6.3
factory-boy==2.12.0
